coffee price funcs bigdecimal

+++ src/main/java/coffeeshop/Controllers/Customers/CustomerNewOrder.java
@@ -2,6 +2,7 @@ package coffeeshop.Controllers.Customers;
 
 import java.io.IOException;
 import java.lang.reflect.Field;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
@@ -45,7 +46,7 @@ public class CustomerNewOrder {
 	@FXML
 	private ComboBox<String> brandBox, roastBox;
 	@FXML
-	private ComboBox<Float> priceBox;
+	private ComboBox<BigDecimal> priceBox;
 	@FXML
 	private ComboBox<Integer> sizeBox;
 	@FXML
@@ -54,7 +55,7 @@ public class CustomerNewOrder {
 	private ObservableList<Coffee> orderList = FXCollections.observableArrayList();
 	private ObservableList<Coffee> filteredList = FXCollections.observableArrayList();
 	private static final List<String> brandsList = CoffeeDAOService.getBrands();
-	private static final List<Float> pricesList = CoffeeDAOService.getPrices();
+	private static final List<BigDecimal> pricesList = CoffeeDAOService.getPrices();
 	private static final List<Integer> sizeList = CoffeeDAOService.getSizes();
 	@FXML
 	private TableColumn<Coffee, String> brandColumn, nameColumn, roastColumn, priceColumn, regionColumn, sizeColumn;
